Add pending requests provider split request method into two
