+++ Sources/Auth/Services/Wallet/PendingRequestsProvider.swift
@@ -1,0 +1,22 @@
+import Foundation
+import JSONRPC
+import WalletConnectUtils
+
+class PendingRequestsProvider {
+    private let rpcHistory: RPCHistory
+
+    init(rpcHistory: RPCHistory) {
+        self.rpcHistory = rpcHistory
+    }
+
+    public func getPendingRequests(account: Account) throws -> [AuthRequest] {
+        let pendingRequests: [AuthRequest] = rpcHistory.getPending()
+            .filter {$0.request.method == "wc_authRequest"}
+            .compactMap {
+                guard let params = try? $0.request.params?.get(AuthRequestParams.self) else {return nil}
+                let message = SIWEMessageFormatter().formatMessage(from: params.payloadParams, address: account.address)
+                return AuthRequest(id: $0.request.id!, message: message)
+            }
+        return pendingRequests
+    }
+}
